Allow remote calls
As requested by @LiddleDev on gitter, this commit allows disabling the
integrated security feature of only allowing localhost to connect to the
ark-rpc server.

This behavior has to explicitly be enabled by supplying `--allow-remote`
as a commandline argument to the ark-rpc server.

The README has been changed to reflect this change.